# docs/apifox/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "网上商城系统 API",
    "description": "Spring Boot 2.7 + MyBatis-Plus + MySQL + Redis + JWT。管理端前缀 /admin，用户端前缀 /user；除登录接口外均需在 Header 携带 token。本文件由 tools/export_openapi.py 从 Controller 源码自动提取生成，与实现同源。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地环境"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "token": {
        "type": "apiKey",
        "in": "header",
        "name": "token",
        "description": "登录接口返回的 JWT，放在请求头 token 字段"
      }
    },
    "schemas": {
...
      "OrdersSubmitDTO": {
        "type": "object",
        "required": [
          "consignee",
          "phone",
          "address",
          "cartItemIds"
        ],
        "properties": {
          "consignee": {
            "type": "string",
            "example": "张三"
          },
          "phone": {
            "type": "string",
            "example": "13900000001"
          },
          "address": {
            "type": "string",
            "example": "上海市嘉定区1号"
          },
          "remark": {
            "type": "string"
          },
          "payMethod": {
            "type": "integer",
            "example": 1
          },
          "cartItemIds": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "example": [
              1,
              2
            ]
          }
        }
      },
...
      "SetmealDTO": {
        "type": "object",
        "required": [
          "name",
          "categoryId",
          "price"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "example": "双人组合套装A"
          },
          "categoryId": {
            "type": "integer",
            "example": 6
          },
          "price": {
            "type": "number",
            "example": 68.0
          },
          "image": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "example": 1
          },
          "setmealDishes": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "token": []
    }
  ],
  "paths": {
...
    "/admin/setmeal": {
      "post": {
        "tags": [
          "组合套装管理"
        ],
        "summary": "save",
        "operationId": "SetmealController_save",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetmealDTO"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "组合套装管理"
        ],
        "summary": "update",
        "operationId": "SetmealController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetmealDTO"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "组合套装管理"
        ],
        "summary": "deleteBatch",
        "operationId": "SetmealController_deleteBatch",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/setmeal/page": {
      "get": {
        "tags": [
          "组合套装管理"
        ],
        "summary": "page",
        "operationId": "SetmealController_page",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/setmeal/status/{status}": {
      "post": {
        "tags": [
          "组合套装管理"
        ],
        "summary": "startOrStop",
        "operationId": "SetmealController_startOrStop",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/setmeal/{id}": {
      "get": {
        "tags": [
          "组合套装管理"
        ],
        "summary": "getById",
        "operationId": "SetmealController_getById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/user/order/cancel/{id}": {
      "put": {
        "tags": [
          "用户订单"
        ],
        "summary": "cancel",
        "operationId": "UserOrderController_cancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/order/history": {
      "get": {
        "tags": [
          "用户订单"
        ],
        "summary": "history",
        "operationId": "UserOrderController_history",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/order/payment/{orderNumber}": {
      "put": {
        "tags": [
          "用户订单"
        ],
        "summary": "paySuccess",
        "operationId": "UserOrderController_paySuccess",
        "parameters": [
          {
            "name": "orderNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/order/submit": {
      "post": {
        "tags": [
          "用户订单"
        ],
        "summary": "submit",
        "operationId": "UserOrderController_submit",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrdersSubmitDTO"
              }
            }
          }
        }
      }
    },
    "/user/order/{id}": {
      "get": {
        "tags": [
          "用户订单"
        ],
        "summary": "details",
        "operationId": "UserOrderController_details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/user/setmeal/{id}": {
      "get": {
        "tags": [
          "商品浏览"
        ],
        "summary": "getById",
        "operationId": "UserDishController_getById",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}